pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[tools]>=0.201.0",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.30.1",
    "langchain-groq>=0.3.8",
    "python-dotenv>=1.1.1",
]
